Extract camera-profile visualization into engine/src/cameraProfile.ts
Moves handleCameraProfile, createCameraVisualization,
createCameraBodyGeometry, createCameraUpArrow, createCameraLabel,
toggleCameraVisibility, updateCameraButtonState,
toggleCameraProfileLabels, toggleCameraProfileCoordinates, and
applyCameraScale out of PointCloudVisualizer - the pyramid+label
visualization built for camera-profile JSON files (distinct from the
pose/keypoint JSON handling in pose.ts). The geometry builders were
already pure; handleCameraProfile and the toggle/scale functions take
a CameraProfileHost parameter.

Also dropped createDirectionArrow, which had zero call sites anywhere
in the codebase (verified via repo-wide grep before removing).

main.ts: 13,053 -> 12,682 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
